backend/ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
scikit-learn==1.5.2
tensorflow==2.16.1
keras==3.0.5
pillow==11.0.0
opencv-python==4.9.0.80
pandas==2.2.0
joblib==1.4.2
python-multipart==0.0.6

